fix incorrect device list when open port with putty on Windows
When any port of FTDI is opened, FT_get_device_info_detail() cannot get
the information of this port.
When the port A information cannot be obtained, try to use the port of
the B/C/D to infer the id of the port A

So here is a limitation. If open all 4 ports are opened with putty, BCU
cannot get the board id.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>