add imx93evk11b1 SR_ pins change support
imx93evk11b1 refers to the combination of imx93 BB revB1 + SOM revB2 and later.
Because the circuit for power consumption measurement is on the SOM board.
Therefore, please use imx93evk11b1 as the BOARD NAME only if the SOM board
is rev B2 and later.

Starting from SOM revB2, a different chip is used to control the size of the
measuring resistance. The function of its OE is just the opposite of the
previous version.

In order to ensure that various combinations of BB and SOM can be used for
normal range switching, add the pin with SRD_ prefix to operate two pins at
the same time.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>