Automatically hide rails with a detection value of 0.
When a rail is not in use, the corresponding current value is set to -2500mA,
at which point these abnormal values are forcibly set to 0.

When using the -dump option, this particular rail will still be recorded
in the dump file, with all its values set to 0.

When displayed on the monitor, rails with an average power of 0 will not
be shown, thereby reducing confusion.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>