all BCU settings will be cleared by default after using the reset command
Based on the concept of "remote debugging", a normal reset command,
once executed, should not affect the subsequent manual local operations
of the board.
Therefore, by default, all BCU outputs are cleared after the reset
is executed.

This ensures that the next time the ON/OFF switch on the board is
manually toggled, it will boot from the BOOT SWITCH on the board.

If you wish to retain the BCU settings, you can use the [-keep] option.

Signed-off-by: Clark Wang <xiaoning.wang@nxp.com>